Dev fanb --》 dev

+++ main.go
@@ -54,14 +54,17 @@ func main() {
 					zap.S().Errorln("获取all user失败:", err)
 				}
 				for _, user := range users {
-					// 如果过期时间在现在之前，则删除
-					if user.ExpiredTime.Before(time.Now()) {
-						// 删除user
-						_, err := service.DeleteUSer(user.ID)
-						if err != nil {
-							zap.S().Errorln("删除user失败:", err)
-						} else {
-							zap.S().Infoln("delete user:", user.Username)
+					// 如果为nil就是永久时长
+					if user.ExpiredTime != nil {
+						// 如果过期时间在现在之前，则删除
+						if user.ExpiredTime.Before(time.Now()) {
+							// 删除user
+							_, err := service.DeleteUSer(user.ID)
+							if err != nil {
+								zap.S().Errorln("删除user失败:", err)
+							} else {
+								zap.S().Infoln("delete user:", user.Username)
+							}
 						}
 					}
 				}
